Combine CardDSL with CardView
